fix: details request in progress modal the change was already made
